addressed comments #235
